+++ base/generator.jl
@@ -52,6 +52,28 @@ size(g::Generator) = size(g.iter)
 axes(g::Generator) = axes(g.iter)
 ndims(g::Generator) = ndims(g.iter)
 
+function getindex(g::Generator{<:AbstractArray}, I...)
+    I′ = to_indices(g.iter, I)
+    subset = g.iter[I′...]
+    if isempty(index_shape(I′...))
+        g.f(subset)
+    else
+        map(g.f, subset)
+    end
+end
+
+function getindex(g::Generator, I...)
+    A = collect(g.iter)
+    subset = A[I...]
+    if typeof(subset) == eltype(A)
+        g.f(subset)
+    else
+        map(g.f, subset)
+    end
+end
+
+firstindex(g::Generator) = firstindex(g.iter)
+lastindex(g::Generator) = lastindex(g.iter)
 
 ## iterator traits